symphonia_core

Module audio

The audio module provides primitives for working with multi-channel audio generically across sample formats.

Standard Sample Formats

The structs and traits in this module are generic across any type implementing the Sample trait. However, a set of standard sample formats is defined. These standard sample formats are: U8, S8, U16, S16, U24, S24, U32, S32, F32, and F64. These sample formats correspond to the following underlying data types, respectively: u8, i8, u16, i16, u24, i24, u32, i32, f32, and f64.

Generic Wrappers

To make handling varying sample formats easier, this module also implements a set of generic wrappers around typed audio buffers and slices. These wrappers are implemented as enums whereby each enumerator corresponds to one of the aforementioned standard sample formats. Functions on the generic wrappers match closely with that of the typed interface and are dispatch to the active enumerator.

Modules

conv

The conv module provides methods to convert samples between different sample types (formats).

layouts

Constants for well known, common, or standardized channel layouts.

sample

The sample module defines the audio sample trait and implements any non-primitive sample data types.

Structs

AudioBuffer

A dynamically sized container for multi-channel planar audio for a known sample format.

AudioPlanes

Iterator over immutable audio plane slices.

AudioPlanesMut

Iterator over mutable audio plane slices.

AudioSlice

An immutable slice of planar audio.

AudioSliceMut

A mutable slice of planar audio.

AudioSpec

A specification defining the core characteristics of some audio.

Interleaved

Iterator over interleaved samples.

Position

A bitmask representing positional audio channels.

Enums

AmbisonicBFormat

Ambisonic B-format channel components up to the 3rd order.

ChannelLabel

A channel label.

Channels

A set of channels.

GenericAudioBuffer

An owning wrapper for an AudioBuffer of any standard sample format.

GenericAudioBufferRef

A non-owning immutable reference wrapper for an AudioBuffer of any standard sample format.

GenericAudioSlice

An owning wrapper for an AudioSlice of any standard sample format.

Traits

AsGenericAudioBufferRef

A trait for generically borrowing an AudioBuffer by wrapping it in a GenericAudioBufferRef.

Audio

Trait for accessing immutable planar audio.

AudioBufferBytes

Trait for querying the maximum capacity in bytes for dynamically sized audio storage.

AudioBytes

Trait for marshalling planar audio to byte buffers.

AudioMut

Trait for manipulating mutable planar audio.