symphonia_core::audio

Enum GenericAudioBuffer

pub enum GenericAudioBuffer {
    U8(AudioBuffer<u8>),
    U16(AudioBuffer<u16>),
    U24(AudioBuffer<u24>),
    U32(AudioBuffer<u32>),
    S8(AudioBuffer<i8>),
    S16(AudioBuffer<i16>),
    S24(AudioBuffer<i24>),
    S32(AudioBuffer<i32>),
    F32(AudioBuffer<f32>),
    F64(AudioBuffer<f64>),
}

An owning wrapper for an AudioBuffer of any standard sample format.

Calls on this wrapper are dispatched to the underlying, wrapped, buffer and are semantically identical.

Variants

U8(AudioBuffer<u8>)

An unsigned 8-bit integer audio buffer.

U16(AudioBuffer<u16>)

An unsigned 16-bit integer audio buffer.

U24(AudioBuffer<u24>)

An unsigned 24-bit integer audio buffer.

U32(AudioBuffer<u32>)

An unsigned 32-bit integer audio buffer.

S8(AudioBuffer<i8>)

A signed 8-bit integer audio buffer.

S16(AudioBuffer<i16>)

A signed 16-bit integer audio buffer.

S24(AudioBuffer<i24>)

A signed 24-bit integer audio buffer.

S32(AudioBuffer<i32>)

A signed 32-bit integer audio buffer.

F32(AudioBuffer<f32>)

A single precision (32-bit) floating point audio buffer.

F64(AudioBuffer<f64>)

A double precision (64-bit) floating point audio buffer.

Implementations

impl GenericAudioBuffer

pub fn new(format: SampleFormat, spec: AudioSpec, capacity: usize) -> Self

pub fn spec(&self) -> &AudioSpec

Get the audio specification.

pub fn num_planes(&self) -> usize

Get the total number of audio planes.

pub fn is_empty(&self) -> bool

Returns true if there are no audio frames.

pub fn frames(&self) -> usize

Gets the number of audio frames in the buffer.

pub fn is_unused(&self) -> bool

Returns true if the AudioBuffer is unused.

An unused AudioBuffer has either a capacity of 0, or no channels.

pub fn capacity(&self) -> usize

Gets the total capacity of the buffer. The capacity is the maximum number of audio frames the buffer can store.

pub fn clear(&mut self)

Clears all audio frames.

pub fn grow_capacity(&mut self, new_capacity: usize)

Grows the capacity of the buffer if the new capacity is larger than the current capacity.

Realtime Safety

This function will allocate if new_capacity exceeds the current capacity.

pub fn resize_with_silence(&mut self, new_len: usize)

Resizes the buffer such that the number of frames is new_len. New frames are silent.

See AudioBuffer::resize_with_silence for full details.

pub fn resize_uninit(&mut self, new_len: usize)

Resizes the buffer such that the number of frames is new_len. New frames are left uninitialized and may contain stale data that should be overwritten.

See AudioBuffer::resize_uninit for full details.

pub fn render_silence(&mut self, num_frames: Option<usize>) -> usize

Renders a number of silent frames.

See AudioBuffer::render_silence for full details.

pub fn render_uninit(&mut self, num_frames: Option<usize>) -> usize

Renders an uninitialized number of frames.

See AudioBuffer::render_uninit for full details.

pub fn shift(&mut self, shift: usize)

Shifts the contents of the buffer back by the number of frames specified. The leading frames are dropped from the buffer.

pub fn slice<R: RangeBounds<usize>>(&self, range: R) -> GenericAudioSlice<'_>

Get an immutable slice of the buffer over range.

pub fn truncate(&mut self, num_frames: usize)

Truncates the buffer to the number of frames specified. If the number of frames in the buffer is less-than the number of frames specified, then this function does nothing.

pub fn trim(&mut self, start: usize, end: usize)

Trims frames from the start and end of the buffer.

pub fn samples_interleaved(&self) -> usize

Get the total number of samples contained in all audio planes.

pub fn samples_planar(&self) -> usize

Get the total number of samples contained in each audio plane.

pub fn copy_to<Sout, Dst>(&self, dst: &mut Dst)
where Sout: ConvertibleSample, Dst: AudioMut<Sout>,

Copy audio to a mutable audio slice while doing any necessary sample format conversions.

pub fn copy_to_slice_interleaved<Sout, Dst>(&self, dst: Dst)
where Sout: ConvertibleSample, Dst: AsMut<[Sout]>,

Copy all audio frames to a slice of samples in interleaved order.

See AudioBuffer::copy_to_slice_interleaved for full details.

pub fn copy_to_slice_planar<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: ConvertibleSample, Dst: AsMut<[Sout]>,

Copy all audio planes to discrete slices.

See AudioBuffer::copy_to_slice_planar for full details.

pub fn copy_to_vec_interleaved<Sout>(&self, dst: &mut Vec<Sout>)
where Sout: ConvertibleSample,

Copy all audio frames to a vector of samples in interleaved order.

See AudioBuffer::copy_to_vec_interleaved for full details.

pub fn copy_to_vecs_planar<Sout>(&self, dst: &mut Vec<Vec<Sout>>)
where Sout: ConvertibleSample,

Copy all audio planes to discrete vectors.

See AudioBuffer::copy_to_vecs_planar for full details.

pub fn copy_bytes_interleaved_as<Sout, Dst>(&self, dst: Dst)
where Sout: SampleBytes + ConvertibleSample, Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice after converting to a different sample format.

See AudioBuffer::copy_bytes_interleaved_as for full details.

pub fn copy_bytes_planar_as<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: SampleBytes + ConvertibleSample, Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane after converting to a different sample format.

See AudioBuffer::copy_bytes_planar_as for full details.

pub fn copy_bytes_interleaved<Dst>(&self, dst: Dst)
where Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice.

See AudioBuffer::copy_bytes_interleaved for full details.

pub fn copy_bytes_planar<Dst>(&self, dst: &mut [Dst])
where Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane.

See AudioBuffer::copy_bytes_planar for full details.

pub fn copy_bytes_to_vec_interleaved(&self, dst: &mut Vec<u8>)

Copy interleaved audio to the destination byte vector.

See AudioBuffer::copy_bytes_to_vec_interleaved for full details.

pub fn copy_bytes_to_vec_interleaved_as<Sout>(&self, dst: &mut Vec<u8>)
where Sout: SampleBytes + ConvertibleSample,

Copy interleaved audio to the destination byte vector after converting to a different sample format.

See AudioBuffer::copy_bytes_to_vec_interleaved_as for full details.

pub fn copy_bytes_to_vecs_planar(&self, dst: &mut Vec<Vec<u8>>)

Copy audio planes as bytes to discrete byte vectors.

See AudioBuffer::copy_bytes_to_vecs_planar for full details.

pub fn copy_bytes_to_vecs_planar_as<Sout>(&self, dst: &mut Vec<Vec<u8>>)
where Sout: SampleBytes + ConvertibleSample,

Copy audio planes as bytes to discrete byte vectors after converting to a different sample format.

See AudioBuffer::copy_bytes_to_vecs_planar_as for full details.

pub fn byte_len_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length in bytes of all samples if converted to a new sample format.

pub fn byte_len_per_plane_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length in bytes of all samples in a single plane if converted to a new sample format.

pub fn byte_len_per_frame_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length of bytes of a single interleaved audio frame if converted to a new sample format.

pub fn byte_len(&self) -> usize

Get the length in bytes of all samples.

pub fn byte_len_per_plane(&self) -> usize

Get the length in bytes of all samples in a single plane.

pub fn byte_len_per_frame(&self) -> usize

Get the length of bytes of a single interleaved audio frame.