symphonia_core::audio

Enum GenericAudioSlice

pub enum GenericAudioSlice<'a> {
    U8(AudioSlice<'a, u8>),
    U16(AudioSlice<'a, u16>),
    U24(AudioSlice<'a, u24>),
    U32(AudioSlice<'a, u32>),
    S8(AudioSlice<'a, i8>),
    S16(AudioSlice<'a, i16>),
    S24(AudioSlice<'a, i24>),
    S32(AudioSlice<'a, i32>),
    F32(AudioSlice<'a, f32>),
    F64(AudioSlice<'a, f64>),
}

An owning wrapper for an AudioSlice of any standard sample format.

Calls on this wrapper are dispatched to the underlying, wrapped, slice and are semantically identical.

Variants

U8(AudioSlice<'a, u8>)

An unsigned 8-bit integer audio slice.

U16(AudioSlice<'a, u16>)

An unsigned 16-bit integer audio slice.

U24(AudioSlice<'a, u24>)

An unsigned 24-bit integer audio slice.

U32(AudioSlice<'a, u32>)

An unsigned 32-bit integer audio slice.

S8(AudioSlice<'a, i8>)

A signed 8-bit integer audio slice.

S16(AudioSlice<'a, i16>)

A signed 16-bit integer audio slice.

S24(AudioSlice<'a, i24>)

A signed 24-bit integer audio slice.

S32(AudioSlice<'a, i32>)

A signed 32-bit integer audio slice.

F32(AudioSlice<'a, f32>)

A single precision (32-bit) floating point audio slice.

F64(AudioSlice<'a, f64>)

A double precision (64-bit) floating point audio slice.

Implementations

impl<'a> GenericAudioSlice<'a>

pub fn spec(&self) -> &AudioSpec

Get the audio specification.

pub fn num_planes(&self) -> usize

Get the total number of audio planes.

pub fn is_empty(&self) -> bool

Returns true if there are no audio frames.

pub fn frames(&self) -> usize

Gets the number of audio frames in the slice.

pub fn samples_interleaved(&self) -> usize

Get the total number of samples contained in all audio planes.

pub fn samples_planar(&self) -> usize

Get the total number of samples contained in each audio plane.

pub fn copy_to<Sout, Dst>(&self, dst: &mut Dst)
where Sout: ConvertibleSample, Dst: AudioMut<Sout>,

Copy audio to a mutable audio slice while doing any necessary sample format conversions.

pub fn copy_to_slice_interleaved<Sout, Dst>(&self, dst: Dst)
where Sout: ConvertibleSample, Dst: AsMut<[Sout]>,

Copy all audio frames to a slice of samples in interleaved order.

See AudioBuffer::copy_to_slice_interleaved for full details.

pub fn copy_to_slice_planar<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: ConvertibleSample, Dst: AsMut<[Sout]>,

Copy all audio planes to discrete slices.

See AudioBuffer::copy_to_slice_planar for full details.

pub fn copy_to_vec_interleaved<Sout>(&self, dst: &mut Vec<Sout>)
where Sout: ConvertibleSample,

Copy all audio frames to a vector of samples in interleaved order.

See AudioBuffer::copy_to_vec_interleaved for full details.

pub fn copy_to_vecs_planar<Sout>(&self, dst: &mut Vec<Vec<Sout>>)
where Sout: ConvertibleSample,

Copy all audio planes to discrete vectors.

See AudioBuffer::copy_to_vecs_planar for full details.

pub fn copy_bytes_interleaved_as<Sout, Dst>(&self, dst: Dst)
where Sout: SampleBytes + ConvertibleSample, Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice after converting to a different sample format.

See AudioBuffer::copy_bytes_interleaved_as for full details.

pub fn copy_bytes_planar_as<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: SampleBytes + ConvertibleSample, Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane after converting to a different sample format.

See AudioBuffer::copy_bytes_planar_as for full details.

pub fn copy_bytes_interleaved<Dst>(&self, dst: Dst)
where Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice.

See AudioBuffer::copy_bytes_interleaved for full details.

pub fn copy_bytes_planar<Dst>(&self, dst: &mut [Dst])
where Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane.

See AudioBuffer::copy_bytes_planar for full details.

pub fn copy_bytes_to_vec_interleaved(&self, dst: &mut Vec<u8>)

Copy interleaved audio to the destination byte vector.

See AudioBuffer::copy_bytes_to_vec_interleaved for full details.

pub fn copy_bytes_to_vec_interleaved_as<Sout>(&self, dst: &mut Vec<u8>)
where Sout: SampleBytes + ConvertibleSample,

Copy interleaved audio to the destination byte vector after converting to a different sample format.

See AudioBuffer::copy_bytes_to_vec_interleaved_as for full details.

pub fn copy_bytes_to_vecs_planar(&self, dst: &mut Vec<Vec<u8>>)

Copy audio planes as bytes to discrete byte vectors.

See AudioBuffer::copy_bytes_to_vecs_planar for full details.

pub fn copy_bytes_to_vecs_planar_as<Sout>(&self, dst: &mut Vec<Vec<u8>>)
where Sout: SampleBytes + ConvertibleSample,

Copy audio planes as bytes to discrete byte vectors after converting to a different sample format.

See AudioBuffer::copy_bytes_to_vecs_planar_as for full details.

pub fn byte_len_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length in bytes of all samples if converted to a new sample format.

pub fn byte_len_per_plane_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length in bytes of all samples in a single plane if converted to a new sample format.

pub fn byte_len_per_frame_as<Sout>(&self) -> usize
where Sout: SampleBytes + ConvertibleSample,

Get the length of bytes of a single interleaved audio frame if converted to a new sample format.

pub fn byte_len(&self) -> usize

Get the length in bytes of all samples.

pub fn byte_len_per_plane(&self) -> usize

Get the length in bytes of all samples in a single plane.

pub fn byte_len_per_frame(&self) -> usize

Get the length of bytes of a single interleaved audio frame.