symphonia_core::audio

Module layouts

Constants for well known, common, or standardized channel layouts.

The channel layout constants defined in this module are defined as Channels::Positioned variants.

Naming Conventions

FRONT layouts include the front left, front left-of-center, front right-of-center, and front right channels.

WIDE layouts include all channels from the FRONT layout plus the front center channel.

SIDE layouts include the side left and side right channels.

XPY layouts have X standard (non-LFE) channels, and Y LFE channels.

Channel Order

Many standardized channel layouts only differ in the order in which the samples for each channel appear in an audio frame (interleaved audio), or the order of the audio planes of each channel (planar audio). These different orderings cannot be represented uniquely by the Position bitmask. Therefore, all such channel layouts will map to the same value.

The order of the channels listed in the detailed description of each channel layout are the order in which the standard describes but is not the order in which Symphonia will store those channels in a buffer.

Constants

CHANNEL_LAYOUT_2P1

Stereophonic audio with low-frequency effects.

CHANNEL_LAYOUT_3P0

3.0 audio.

CHANNEL_LAYOUT_3P0_REAR

3.0 audio with rear surround channel.

CHANNEL_LAYOUT_3P1

3.1 audio.

CHANNEL_LAYOUT_4P0

4.0 audio.

CHANNEL_LAYOUT_4P0_QUAD

Quadrophonic audio with rear channels.

CHANNEL_LAYOUT_4P0_QUAD_SIDE

Quadrophonic audio with side channels.

CHANNEL_LAYOUT_4P1

4.1 audio.

CHANNEL_LAYOUT_5P0

5.0 audio.

CHANNEL_LAYOUT_5P0_SIDE

5.0 audio with side channels.

CHANNEL_LAYOUT_5P1

5.1 audio.

CHANNEL_LAYOUT_5P1_SIDE

5.1 audio with side channels and low-frequency effects.

CHANNEL_LAYOUT_6P0

6.0 audio.

CHANNEL_LAYOUT_6P0_FRONT

6.0 audio with additional front channels.

CHANNEL_LAYOUT_6P0_HEX

Hexagonal audio.

CHANNEL_LAYOUT_6P1

6.1 audio with low-frequency effects.

CHANNEL_LAYOUT_6P1_FRONT

6.1 audio with additional front channels and low-frequency effects.

CHANNEL_LAYOUT_6P1_HEX

Hexagonal audio with low-frequency effects.

CHANNEL_LAYOUT_7P0

7.0 surround audio.

CHANNEL_LAYOUT_7P0_WIDE

7.0 audio with wide front channels.

CHANNEL_LAYOUT_7P0_WIDE_SIDE

7.0 audio with wide front channels and side channels.

CHANNEL_LAYOUT_7P1

7.1 surround audio with low-frequency effects.

CHANNEL_LAYOUT_7P1_WIDE

7.1 audio with wide front channels and low-frequency effects.

CHANNEL_LAYOUT_7P1_WIDE_SIDE

7.1 audio with wide front channels and side channels.

CHANNEL_LAYOUT_22P2

22.2 audio.

CHANNEL_LAYOUT_AAC_3P0

AAC 3-channel channel layout.

CHANNEL_LAYOUT_AAC_4P0

AAC 4-channel surround-based layout.

CHANNEL_LAYOUT_AAC_5P0

AAC 5-channel surround-based layout.

CHANNEL_LAYOUT_AAC_5P1

AAC 5.1-channel surround-based layout.

CHANNEL_LAYOUT_AAC_6P0

AAC 6-channel surround-based layout.

CHANNEL_LAYOUT_AAC_6P1

AAC 6.1-channel surround-based layout.

CHANNEL_LAYOUT_AAC_7P1

AAC 7.1-channel surround-based layout.

CHANNEL_LAYOUT_AAC_QUADRAPHONIC

AAC quadraphonic surround-based layout.

CHANNEL_LAYOUT_MONO

Monophonic audio.

CHANNEL_LAYOUT_MPEG_1P0

MPEG 1-channel channel layout.

CHANNEL_LAYOUT_MPEG_2P0

MPEG 2-channel channel layout.

CHANNEL_LAYOUT_MPEG_3P0_A

MPEG 3-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_3P0_B

MPEG 3-channel, configuration B, channel layout.

CHANNEL_LAYOUT_MPEG_4P0_A

MPEG 4-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_4P0_B

MPEG 4-channel, configuration B, channel layout.

CHANNEL_LAYOUT_MPEG_5P0_A

MPEG 5-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_5P0_B

MPEG 5-channel, configuration B, channel layout.

CHANNEL_LAYOUT_MPEG_5P0_C

MPEG 5-channel, configuration C, channel layout.

CHANNEL_LAYOUT_MPEG_5P0_D

MPEG 5-channel, configuration D, channel layout.

CHANNEL_LAYOUT_MPEG_5P1_A

MPEG 5.1-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_5P1_B

MPEG 5.1-channel, configuration B, channel layout.

CHANNEL_LAYOUT_MPEG_5P1_C

MPEG 5.1-channel, configuration C, channel layout.

CHANNEL_LAYOUT_MPEG_5P1_D

MPEG 5.1-channel, configuration D, channel layout.

CHANNEL_LAYOUT_MPEG_6P1_A

MPEG 6.1-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_7P1_A

MPEG 7.1-channel, configuration A, channel layout.

CHANNEL_LAYOUT_MPEG_7P1_B

MPEG 7.1-channel, configuration B, channel layout.

CHANNEL_LAYOUT_OGG_1P0

OGG 1-channel layout.

CHANNEL_LAYOUT_OGG_2P0

OGG 2-channel layout.

CHANNEL_LAYOUT_OGG_3P0

OGG 3-channel layout.

CHANNEL_LAYOUT_OGG_4P0

OGG 4-channel layout.

CHANNEL_LAYOUT_OGG_5P0

OGG 5-channel layout.

CHANNEL_LAYOUT_OGG_5P1

OGG 5.1-channel layout.

CHANNEL_LAYOUT_OGG_6P1

OGG 6.1-channel layout.

CHANNEL_LAYOUT_OGG_7P1

OGG 7.1-channel layout.

CHANNEL_LAYOUT_STEREO

Stereophonic audio.