symphonia_core::audio::sample

Trait SampleBytes

pub trait SampleBytes: Sample {
    type RawType: ByteAligned;

    // Required methods
    fn to_ne_sample_bytes(self) -> Self::RawType;
    fn to_le_sample_bytes(self) -> Self::RawType;
    fn to_be_sample_bytes(self) -> Self::RawType;
}

SampleBytes provides interfaces to get the packed byte representation of a Sample.

The packed byte representation is the byte representation of a sample in its native stream format. This may differ from its in-memory working representation.

Required Associated Types

type RawType: ByteAligned

The data type that stores the packed byte representation of the sample.

The sealed::ByteAligned trait bound constrains the allowable types this may be. Only byte array types: [u8; N] for N between 1 and 16 inclusive are allowed.

Required Methods

fn to_ne_sample_bytes(self) -> Self::RawType

Return the sample in its packed byte representation in the native byte order.

fn to_le_sample_bytes(self) -> Self::RawType

Return the sample in its packed byte representation in the little-endian byte order.

fn to_be_sample_bytes(self) -> Self::RawType

Return the sample in its packed byte representation in the big-endian byte order.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.