symphonia_core::audio

Struct AudioBuffer

pub struct AudioBuffer<S: Sample> { /* private fields */ }

A dynamically sized container for multi-channel planar audio for a known sample format.

An AudioBuffer is characterized by a total capacity, and an audio specification describing the audio channels and sample rate. The capacity of an AudioBuffer is the fixed number of audio frames the buffer may store without growing.

Note, AudioBuffer does not automatically grow its capacity. All increases in capacity must be manually requested.

AudioBuffer stores each audio channel in separate audio planes. An audio plane is a continguous vector of audio samples. The samples across all audio planes at any one particular index form an audio frame. The order of audio planes, and thus the order of samples in an audio frame, is a canonical order based on the channels mask.

Audio Plane Ordering

The order and indexing of the audio planes in an audio buffer is based on the channel representation specified in the audio specification.

If Channels::Positioned, the audio buffer contains one audio plane for each bit set in the specified Position bitmask. The order of audio planes is based on the position bits set to 1 from the least-significant bit (LSB) to most-significant bit (MSB) in the mask. This order is known as the canonical channel order. The audio buffer may be indexed by a position mask containing exactly one position, or a canonical index.

If Channels::Discrete, the total number of discrete channels (n) is specified. The audio buffer contains one audio plane for discrete channel indicies 0..n. The buffer may be indexed by a discrete channel index, i, in the range 0..n.

If Channels::Ambisonic, the Ambisonic order (n) is specified. The audio buffer contains m = (1 + n) ^ 2 audio planes, one for each Ambisonic component. The order of audio planes is based on the Ambisonic Channel Number (ACN). The audio buffer may be indexed by the ACN of the desired channel, k, in the range 0..m.

If Channels::Custom, a list of n ChannelLabels is specified. Each channel label describes the disposition of the channel. The audio buffer contains n audio planes in the same order as specified in the list. The audio buffer may be indexed by an index, i, in the range 0..n.

Realtime Safety

Unless explictly noted, AudioBuffer never allocates after instantiation.

Implementations

impl<S: Sample> AudioBuffer<S>

pub fn new(spec: AudioSpec, capacity: usize) -> Self

Instantiate a new AudioBuffer using the specified signal specification and an initial capacity in number of samples.

pub fn is_unused(&self) -> bool

Returns true if the AudioBuffer is unused.

An unused AudioBuffer has either a capacity of 0, or no channels.

pub fn capacity(&self) -> usize

Gets the total capacity of the buffer. The capacity is the maximum number of audio frames the buffer can store.

pub fn clear(&mut self)

Clears all audio frames.

pub fn grow_capacity(&mut self, new_capacity: usize)

Grows the capacity of the buffer if the new capacity is larger than the current capacity.

Realtime Safety

This function will allocate if new_capacity exceeds the current capacity.

pub fn slice<R: RangeBounds<usize>>(&self, range: R) -> AudioSlice<'_, S>

Get an immutable slice of the buffer over range.

pub fn slice_mut<R: RangeBounds<usize>>( &mut self, range: R, ) -> AudioSliceMut<'_, S>

Get a mutable slice of the buffer over range.

pub fn append<Sin, Src>(&mut self, src: &Src)
where Sin: Sample, S: Sample + FromSample<Sin>, Src: Audio<Sin>,

Append the frames from the audio buffer src to this buffer.

Panics

Panics if src does not have the same specification, or the combined length would exceed the capacity of this buffer.

pub fn resize(&mut self, new_len: usize, frame: &[S])

Resizes the buffer such that the number of frames is new_len. New frames are copies of frame.

This function does not modify the maximum capacity of the buffer. If new_len is larger than the current number of frames in the buffer, the buffer will be padded with copies of frame. If new_len is smaller than the current number of frames, the buffer will be truncated.

Panics

Panics if new_len exceeds the buffer capacity.

pub fn resize_with_sample(&mut self, new_len: usize, sample: S)

Resizes the buffer such that the number of frames is new_len. New frames repeat sample for all planes.

This function does not modify the maximum capacity of the buffer. If new_len is larger than the current number of frames in the buffer, the buffer will be padded with sample. If new_len is smaller than the current number of frames, the buffer will be truncated.

Panics

Panics if new_len exceeds the buffer capacity.

pub fn resize_with_silence(&mut self, new_len: usize)

Resizes the buffer such that the number of frames is new_len. New frames are silent.

This function does not modify the maximum capacity of the buffer. If new_len is larger than the current number of frames in the buffer, then the buffer will be padded with silent frames. If new_len is smaller than the current number of frames, the buffer will be truncated.

Panics

Panics if new_len exceeds the buffer capacity.

pub fn resize_uninit(&mut self, new_len: usize)

Resizes the buffer such that the number of frames is new_len. New frames are left uninitialized and may contain stale data that should be overwritten.

This function does not modify the maximum capacity of the buffer.

Panics

Panics if new_len exceeds the buffer capacity.

pub fn render(&mut self, num_frames: Option<usize>, frame: &[S]) -> usize

Renders a number of repeated frames.

If num_frames is None, the remaining capacity will be used.

Returns the number of frames rendered.

Panics

Panics if num_frames is Some and would exceed the capacity of the buffer, or frame is not the same length as the number of planes.

pub fn render_with<F>( &mut self, num_frames: Option<usize>, f: F, ) -> Result<usize>
where F: FnMut(usize, &mut [&mut [S]]) -> Result<()>,

Renders a number of frames using the provided render function.

The number of frames to render is specified by num_frames. If num_frames is None, the remaining capacity will be used.

The render function, f, is called once per frame and is expected to populate the next sample in each audio plane. The slice of samples for each audio plane cover all existing samples and the space for new samples. The index of the next sample to write is provided to the call. All samples beyond the write index should be considered uninitialized with stale data. The planes are in the canonical order. If the render function returns an error, the render operation is terminated and the error is returned. Only the samples written upto that point are committed.

On success, the number of frames rendered is returned.

Realtime Safety

If the number of planes exceeds 8, this function will allocate.

Panics

Panics if num_frames is Some and would exceed the capacity of the buffer.

pub fn render_silence(&mut self, num_frames: Option<usize>) -> usize

Renders a number of silent frames.

If num_frames is None, the remaining capacity will be used.

Returns the number of frames rendered.

Panics

Panics if num_frames is Some and would exceed the capacity of the buffer.

pub fn render_uninit(&mut self, num_frames: Option<usize>) -> usize

Renders an uninitialized number of frames.

This is a cheap operation and simply advances the frame counter. The underlying samples are not modified and should be overwritten.

If num_frames is None, the remaining capacity will be used.

Returns the number of frames rendered.

Panics

Panics if num_frames is Some and would exceed the capacity of the buffer.

pub fn shift(&mut self, shift: usize)

Shifts the contents of the buffer back by the number of frames specified. The leading frames are dropped from the buffer.

pub fn truncate(&mut self, num_frames: usize)

Truncates the buffer to the number of frames specified. If the number of frames in the buffer is less-than the number of frames specified, then this function does nothing.

pub fn trim(&mut self, start: usize, end: usize)

Trims frames from the start and end of the buffer.