symphonia_core::audio

Trait Audio

pub trait Audio<S: Sample> {

    // Required methods
    fn spec(&self) -> &AudioSpec;
    fn num_planes(&self) -> usize;
    fn is_empty(&self) -> bool;
    fn frames(&self) -> usize;
    fn plane(&self, idx: usize) -> Option<&[S]>;
    fn plane_pair(&self, idx0: usize, idx1: usize) -> Option<(&[S], &[S])>;
    fn iter_planes(&self) -> AudioPlanes<'_, S> ;
    fn iter_interleaved(&self) -> Interleaved<'_, S> ;
    fn copy_to_slice_interleaved<Sout, Dst>(&self, dst: Dst)
       where Sout: Sample + FromSample<S>,
             Dst: AsMut<[Sout]>;

    // Provided methods
    fn plane_by_position(&self, pos: Position) -> Option<&[S]> { ... }
    fn plane_pair_by_position(
        &self,
        pos0: Position,
        pos1: Position,
    ) -> Option<(&[S], &[S])> { ... }
    fn samples_interleaved(&self) -> usize { ... }
    fn samples_planar(&self) -> usize { ... }
    fn copy_to_slice_planar<Sout, Dst>(&self, dst: &mut [Dst])
       where Sout: Sample + FromSample<S>,
             Dst: AsMut<[Sout]> { ... }
    fn copy_to_vec_interleaved<Sout>(&self, dst: &mut Vec<Sout>)
       where Sout: Sample + FromSample<S> { ... }
    fn copy_to_vecs_planar<Sout>(&self, dst: &mut Vec<Vec<Sout>>)
       where Sout: Sample + FromSample<S> { ... }

}

Trait for accessing immutable planar audio.

Required Methods

fn spec(&self) -> &AudioSpec

Get the audio specification.

fn num_planes(&self) -> usize

Get the total number of audio planes.

fn is_empty(&self) -> bool

Returns true if there are no audio frames.

fn frames(&self) -> usize

Gets the number of audio frames in the buffer.

fn plane(&self, idx: usize) -> Option<&[S]>

Try to get an immutable slice to the audio plane at the canonical buffer index idx.

fn plane_pair(&self, idx0: usize, idx1: usize) -> Option<(&[S], &[S])>

Try to get immutable slices to a pair of audio planes at the canonical buffer indicies idx0 and idx1.

Panics

Panics if idx0 and idx1 are the same index.

fn iter_planes(&self) -> AudioPlanes<'_, S>

Iterate over immutable slices to all planes.

fn iter_interleaved(&self) -> Interleaved<'_, S>

Iterate over all samples in an interleaved order.

fn copy_to_slice_interleaved<Sout, Dst>(&self, dst: Dst)
where Sout: Sample + FromSample<S>, Dst: AsMut<[Sout]>,

Copy all audio frames to a slice of samples in interleaved order.

If the sample format of the slice differs a sample format conversion will occur. Samples are interleaved in canonical order.

Panics

Panics if the length of the destination slice is not the correct length. Use samples_interleaved to get the correct length.

Provided Methods

fn plane_by_position(&self, pos: Position) -> Option<&[S]>

Try to get an immutable slice to the audio plane at position pos.

Panics

Panics if pos contains more than one position.

fn plane_pair_by_position( &self, pos0: Position, pos1: Position, ) -> Option<(&[S], &[S])>

Try to get an immutable slice to a pair of audio planes at positions pos0 and pos1.

Panics

Panics if pos0 and pos1 are the same position, or contain more than one position.

fn samples_interleaved(&self) -> usize

Get the total number of samples contained in all audio planes.

fn samples_planar(&self) -> usize

Get the total number of samples contained in each audio plane.

fn copy_to_slice_planar<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: Sample + FromSample<S>, Dst: AsMut<[Sout]>,

Copy all audio planes to discrete slices.

If the sample format of a slice differs a sampe format conversion will occur.

Panics

Panics if the number of slices is not equal to the number of audio planes, or if a slice is not the correct length. Use num_planes to get the number of audio planes, and samples_planar to get the length of each plane.

fn copy_to_vec_interleaved<Sout>(&self, dst: &mut Vec<Sout>)
where Sout: Sample + FromSample<S>,

Copy all audio frames to a vector of samples in interleaved order.

This function takes a mutable reference to a vector. The vector is resized such that the length of the vector after the copy is the exact number of samples copied.

A sample format conversion will occur if the sample format of the destination vector differs.

Realtime Safety

This function allocates if the vector is not long enough. Allocations can be avoided if the vector has capacity reserved ahead of time.

fn copy_to_vecs_planar<Sout>(&self, dst: &mut Vec<Vec<Sout>>)
where Sout: Sample + FromSample<S>,

Copy all audio planes to discrete vectors.

This function takes a mutable reference to a vector of vectors. The outer vector is resized to match the number of audio planes being copied. Each inner audio plane vector is then resized to match the number of samples copied per plane.

A sample format conversion will occur if the sample format of the destination vectors differ.

Realtime Safety

This function allocates if either the outer vector is not long enough, or if an audio plane vector is not long enough. Allocations can be avoided if all vectors have capacity reserved ahead of time.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.