symphonia_core::audio

Trait AudioBufferBytes

pub trait AudioBufferBytes<S: Sample + SampleBytes>: AudioBytes<S> {
    // Required methods
    fn max_byte_len_per_plane_as<Sout>(&self) -> usize
       where Sout: SampleBytes;
    fn max_byte_len_per_plane(&self) -> usize;

    // Provided methods
    fn max_byte_len_as<Sout>(&self) -> usize
       where Sout: SampleBytes { ... }
    fn max_byte_len(&self) -> usize { ... }
}

Trait for querying the maximum capacity in bytes for dynamically sized audio storage.

Required Methods

fn max_byte_len_per_plane_as<Sout>(&self) -> usize
where Sout: SampleBytes,

Get the maximum possible length in bytes of all samples in a single plane after converting to a different sample format.

fn max_byte_len_per_plane(&self) -> usize

Get the maximum possible length in bytes of all samples in a single plane.

Provided Methods

fn max_byte_len_as<Sout>(&self) -> usize
where Sout: SampleBytes,

Get the maximum possible length in bytes of all samples after converting to a different sample format.

fn max_byte_len(&self) -> usize

Get the maximum possible length in bytes of all samples.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.