symphonia_core::audio

Trait AudioBytes

pub trait AudioBytes<S: Sample + SampleBytes>: Audio<S> {

    // Required methods
    fn copy_bytes_interleaved_as<Sout, Dst>(&self, dst: Dst)
       where Sout: SampleBytes + FromSample<S>,
             Dst: AsMut<[u8]>;
    fn copy_bytes_planar_as<Sout, Dst>(&self, dst: &mut [Dst])
       where Sout: SampleBytes + FromSample<S>,
             Dst: AsMut<[u8]>;
    fn copy_bytes_interleaved<Dst>(&self, dst: Dst)
       where Dst: AsMut<[u8]>;
    fn copy_bytes_planar<Dst>(&self, dst: &mut [Dst])
       where Dst: AsMut<[u8]>;

    // Provided methods
    fn byte_len_as<Sout>(&self) -> usize
       where Sout: SampleBytes + FromSample<S> { ... }
    fn byte_len_per_plane_as<Sout>(&self) -> usize
       where Sout: SampleBytes + FromSample<S> { ... }
    fn byte_len_per_frame_as<Sout>(&self) -> usize
       where Sout: SampleBytes + FromSample<S> { ... }
    fn byte_len(&self) -> usize { ... }
    fn byte_len_per_plane(&self) -> usize { ... }
    fn byte_len_per_frame(&self) -> usize { ... }
    fn copy_bytes_to_vec_interleaved(&self, dst: &mut Vec<u8>) { ... }
    fn copy_bytes_to_vec_interleaved_as<Sout>(&self, dst: &mut Vec<u8>)
       where Sout: SampleBytes + FromSample<S> { ... }
    fn copy_bytes_to_vecs_planar(&self, dst: &mut Vec<Vec<u8>>) { ... }
    fn copy_bytes_to_vecs_planar_as<Sout>(&self, dst: &mut Vec<Vec<u8>>)
       where Sout: SampleBytes + FromSample<S> { ... }

}

Trait for marshalling planar audio to byte buffers.

Required Methods

fn copy_bytes_interleaved_as<Sout, Dst>(&self, dst: Dst)
where Sout: SampleBytes + FromSample<S>, Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice after converting to a different sample format.

The destination slice must be exactly the correct length to fit all audio frames. Use byte_len_as to determine this length.

Panics

Panics if dst is not the expected length.

fn copy_bytes_planar_as<Sout, Dst>(&self, dst: &mut [Dst])
where Sout: SampleBytes + FromSample<S>, Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane after converting to a different sample format.

There must be exactly one destination slice per audio plane.

Each destination slice must be exactly the correct length to fit all audio samples. Use byte_len_per_plane_as to determine this length.

Panics

Panics if:

The length of dst is not exactly equal to the number of planes.

The length of each slice in dst is not the expected length.

fn copy_bytes_interleaved<Dst>(&self, dst: Dst)
where Dst: AsMut<[u8]>,

Copy interleaved audio to the destination byte slice.

The destination slice must be exactly the correct length to fit all audio frames. Use byte_len to determine this length.

Panics

Panics if dst is not the expected length.

fn copy_bytes_planar<Dst>(&self, dst: &mut [Dst])
where Dst: AsMut<[u8]>,

Copy planar audio as bytes to a destination slice per plane.

There must be exactly one destination slice per audio plane.

Each destination slice must be exactly the correct length to fit all audio samples. Use byte_len_per_plane to determine this length.

Panics

Panics if:

The length of dst is not exactly equal to the number of planes.

The length of each slice in dst is not the expected length.

Provided Methods

fn byte_len_as<Sout>(&self) -> usize
where Sout: SampleBytes + FromSample<S>,

Get the length in bytes of all samples if converted to a new sample format.

fn byte_len_per_plane_as<Sout>(&self) -> usize
where Sout: SampleBytes + FromSample<S>,

Get the length in bytes of all samples in a single plane if converted to a new sample format.

fn byte_len_per_frame_as<Sout>(&self) -> usize
where Sout: SampleBytes + FromSample<S>,

Get the length of bytes of a single interleaved audio frame if converted to a new sample format.

fn byte_len(&self) -> usize

Get the length in bytes of all samples.

fn byte_len_per_plane(&self) -> usize

Get the length in bytes of all samples in a single plane.

fn byte_len_per_frame(&self) -> usize

Get the length of bytes of a single interleaved audio frame.

fn copy_bytes_to_vec_interleaved(&self, dst: &mut Vec<u8>)

Copy interleaved audio to the destination byte vector.

This function takes a mutable reference to a vector. The vector is resized such that the length of the vector after the copy is the exact number of samples copied.

Realtime Safety

This function allocates if the vector is not long enough. Allocations can be avoided if the vector has capacity reserved ahead of time.

fn copy_bytes_to_vec_interleaved_as<Sout>(&self, dst: &mut Vec<u8>)
where Sout: SampleBytes + FromSample<S>,

Copy interleaved audio to the destination byte vector after converting to a different sample format.

This function takes a mutable reference to a vector. The vector is resized such that the length of the vector after the copy is the exact number of samples copied.

Realtime Safety

This function allocates if the vector is not long enough. Allocations can be avoided if the vector has capacity reserved ahead of time.

fn copy_bytes_to_vecs_planar(&self, dst: &mut Vec<Vec<u8>>)

Copy audio planes as bytes to discrete byte vectors.

This function takes a mutable reference to a vector of vectors. The outer vector is resized to match the number of audio planes being copied. Each inner byte vector is resized to match the number of bytes per plane.

Realtime Safety

This function allocates if either the outer vector is not long enough, or if an audio plane vector is not long enough. Allocations can be avoided if all vectors have capacity reserved ahead of time.

fn copy_bytes_to_vecs_planar_as<Sout>(&self, dst: &mut Vec<Vec<u8>>)
where Sout: SampleBytes + FromSample<S>,

Copy audio planes as bytes to discrete byte vectors after converting to a different sample format.

This function takes a mutable reference to a vector of vectors. The outer vector is resized to match the number of audio planes being copied. Each inner byte vector is resized to match the number of bytes per plane.

Realtime Safety

This function allocates if either the outer vector is not long enough, or if an audio plane vector is not long enough. Allocations can be avoided if all vectors have capacity reserved ahead of time.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.