symphonia_core::audio

Trait AudioMut

pub trait AudioMut<S: Sample>: Audio<S> {
    // Required methods
    fn plane_mut(&mut self, idx: usize) -> Option<&mut [S]>;
    fn plane_pair_mut(
        &mut self,
        idx0: usize,
        idx1: usize,
    ) -> Option<(&mut [S], &mut [S])>;
    fn iter_planes_mut(&mut self) -> AudioPlanesMut<'_, S> ;
    fn copy_from_slice_interleaved<Sin, Src>(&mut self, src: &Src)
       where Sin: Sample,
             S: Sample + FromSample<Sin>,
             Src: AsRef<[Sin]>;

    // Provided methods
    fn plane_by_position_mut(&mut self, pos: Position) -> Option<&mut [S]> { ... }
    fn plane_pair_by_position_mut(
        &mut self,
        pos0: Position,
        pos1: Position,
    ) -> Option<(&mut [S], &mut [S])> { ... }
    fn apply<F>(&mut self, f: F)
       where F: Fn(S) -> S { ... }
    fn copy_from<Sin, Src>(&mut self, src: &Src)
       where Sin: Sample,
             S: Sample + FromSample<Sin>,
             Src: Audio<Sin> { ... }
    fn copy_from_slice_planar<Sin, Src>(&mut self, src: &[Src])
       where Sin: Sample,
             S: Sample + FromSample<Sin>,
             Src: AsRef<[Sin]> { ... }
}

Trait for manipulating mutable planar audio.

Required Methods

fn plane_mut(&mut self, idx: usize) -> Option<&mut [S]>

Try to get a mutable slice to the audio plane at the canonical buffer index idx.

fn plane_pair_mut( &mut self, idx0: usize, idx1: usize, ) -> Option<(&mut [S], &mut [S])>

Try to get mutable slices to a pair of audio planes at the canonical buffer indicies idx0 and idx1.

Panics

Panics if idx0 and idx1 are the same index.

fn iter_planes_mut(&mut self) -> AudioPlanesMut<'_, S>

Iterate over mutable slices to all planes.

fn copy_from_slice_interleaved<Sin, Src>(&mut self, src: &Src)
where Sin: Sample, S: Sample + FromSample<Sin>, Src: AsRef<[Sin]>,

Copy audio from an interleaved slice of samples.

Provided Methods

fn plane_by_position_mut(&mut self, pos: Position) -> Option<&mut [S]>

Try to get a mutable slice to the audio plane at position pos.

Panics

Panics if pos contains the same position.

fn plane_pair_by_position_mut( &mut self, pos0: Position, pos1: Position, ) -> Option<(&mut [S], &mut [S])>

Try to get mutable slices to a pair of audio planes at positions pos0 and pos1.

Panics

Panics if pos0 and pos1 are the same position, or contains the same position.

fn apply<F>(&mut self, f: F)
where F: Fn(S) -> S,

Applies a transformation function over the samples of all planes.

The order of traversal is not specified.

fn copy_from<Sin, Src>(&mut self, src: &Src)
where Sin: Sample, S: Sample + FromSample<Sin>, Src: Audio<Sin>,

Copy audio from a source.

Panics

Panics if:

The audio specification of this and the source are not identical.

The number of frames of both this and the source are not identical.

fn copy_from_slice_planar<Sin, Src>(&mut self, src: &[Src])
where Sin: Sample, S: Sample + FromSample<Sin>, Src: AsRef<[Sin]>,

Copy audio from a slice of slices.

Panics

Panics if:

The number of slices does not match the number of audio planes.

The number of samples in each slice does not match the number of frames.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.