symphonia_core::codecs

Module audio

Audio decoder specific support.

Modules

well_known

Codec IDs and profiles for well-known audio codecs.

Structs

AudioCodecId

An AudioCodecId is a unique identifier used to identify a specific audio codec.

AudioCodecParameters

Codec parameters for audio codecs.

AudioDecoderOptions

AudioDecoderOptions is a common set of options that all audio decoders use.

FinalizeResult

FinalizeResult contains optional information that can only be found, calculated, or determined after decoding is complete.

Enums

VerificationCheck

A method and expected value to perform verification on the decoded audio.

Constants

CODEC_ID_NULL_AUDIO

Null audio codec ID

Traits

AudioDecoder

An AudioDecoder implements an audio codec’s decode algorithm. It consumes Packets and produces buffers of PCM audio.