symphonia_core::codecs

Module registry

Registry for codecs to support lookup and instantiation of decoders dynamically at runtime.

Structs

CodecRegistry

A CodecRegistry allows the registration of codecs, and provides a method to instantiate a Decoder given a CodecParameters object.

RegisteredAudioDecoder

Registration details of an audio decoder for a particular audio codec.

SupportedAudioCodec

Description of a supported audio codec.

Traits

RegisterableAudioDecoder

To support registration in a codec registry, an AudioDecoder must implement the RegisterableAudioDecoder trait.

Type Aliases

AudioDecoderFactoryFn

AudioDecoder factory function. Creates a boxed AudioDecoder.