symphonia_core::codecs::registry

Struct CodecRegistry

pub struct CodecRegistry { /* private fields */ }

A CodecRegistry allows the registration of codecs, and provides a method to instantiate a Decoder given a CodecParameters object.

Implementations

impl CodecRegistry

pub fn new() -> Self

Instantiate a new CodecRegistry.

pub fn get_audio_decoder( &self, id: AudioCodecId, ) -> Option<&RegisteredAudioDecoder>

Get the registration information of the most preferred audio decoder for the specified audio codec.

pub fn get_audio_decoder_at_tier( &self, tier: Tier, id: AudioCodecId, ) -> Option<&RegisteredAudioDecoder>

Get the registration information of the audio decoder at the specified tier for the specified audio codec.

pub fn register_audio_decoder<C: RegisterableAudioDecoder>(&mut self)

Registers all audio codecs supported by the audio decoder at the standard tier.

If a supported audio codec was previously registered by another audio decoder at the same tier, it will be replaced within the registry.

pub fn register_audio_decoder_at_tier<C: RegisterableAudioDecoder>( &mut self, tier: Tier, )

Registers all audio codecs supported by the audio decoder at a specific tier.

If a supported codec was previously registered by another audio decoder at the same tier, it will be replaced within the registry.

pub fn make_audio_decoder( &self, params: &AudioCodecParameters, opts: &AudioDecoderOptions, ) -> Result<Box<dyn AudioDecoder>>

Instantiate an audio decoder for the specified audio codec parameters.

This function searches the registry for an audio decoder that supports the codec. If one is found, it will be instantiated with the provided audio codec parameters and audio decoder options. If a suitable decoder could not be found, or the decoder could not be instantiated, an error will be returned.