symphonia_core::codecs

Struct CodecRegistry

pub struct CodecRegistry { /* private fields */ }

A CodecRegistry allows the registration of codecs, and provides a method to instantiate a Decoder given a CodecParameters object.

Implementations

impl CodecRegistry

pub fn new() -> Self

Instantiate a new CodecRegistry.

pub fn get_codec(&self, codec: CodecType) -> Option<&CodecDescriptor>

Gets the CodecDescriptor for a registered codec.

pub fn register_all<D: Decoder>(&mut self)

Registers all codecs supported by Decoder. If a supported codec was previously registered by another Decoder it will be replaced within the registry.

pub fn register(&mut self, descriptor: &CodecDescriptor)

Register a single codec. If the codec was previously registered it will be replaced within the registry.

pub fn make( &self, params: &CodecParameters, options: &DecoderOptions, ) -> Result<Box<dyn Decoder>>

Searches the registry for a Decoder that supports the codec. If one is found, it will be instantiated with the provided CodecParameters and returned. If a Decoder could not be found, or the CodecParameters are either insufficient or invalid for the Decoder, an error will be returned.