Module symphonia_core::conv

The conv module provides methods to convert samples between different sample types (formats).

Modules

dither

The dither module provides methods to apply a dither to a sample.

Traits

ConvertibleSample

FromSample

FromSample implements a conversion from Sample type F to Self.

IntoSample

IntoSample implements a conversion from Self to Sample type T.

ReversibleSample

ReversibleSample is a trait that when implemented for Self, that Sample type implements reversible conversions between Self and the parameterized Sample type S.