symphonia_core::conv::dither

Trait Dither

pub trait Dither<F: Sample, T: Sample> {
    // Required method
    fn dither(&mut self, sample: F) -> F;
}

Dither is a trait for implementing dithering algorithms.

Required Methods

fn dither(&mut self, sample: F) -> F

Dithers a Sample of source sample format F for an eventual conversion to the destination sample format T.