Trait symphonia_core::conv::FromSample

pub trait FromSample<F> {
    // Required method
    fn from_sample(val: F) -> Self;
}

FromSample implements a conversion from Sample type F to Self.

This may be a lossy conversion if converting from a sample type of higher precision to one of lower precision. No dithering is applied.

Required Methods

fn from_sample(val: F) -> Self

Object Safety

This trait is not object safe.