symphonia_core::conv

Trait IntoSample

pub trait IntoSample<T> {
    // Required method
    fn into_sample(self) -> T;
}

IntoSample implements a conversion from Self to Sample type T.

This may be a lossy conversion if converting from a sample type of higher precision to one of lower precision. No dithering is applied.

Required Methods

fn into_sample(self) -> T