symphonia_core

Module formats

The format module provides the traits and support structures necessary to implement media demuxers.

Modules

prelude

The formats module prelude for format reader implementers.

probe

The probe module provides methods and traits to support auto-detection of media formats from arbitrary media streams.

util

Helper utilities for implementing FormatReaders.

well_known

IDs for well-known container formats.

Structs

ExternalFormatData

ExternalFormatData contains supplementary data related to the media container that was read before the start of the container, or provided through some other side-channel.

FileAttachment

A file attachment.

FormatId

A FormatId is a unique identifier used to identify a specific container format.

FormatInfo

Basic information about a container format.

FormatOptions

FormatOptions is a common set of options that all demuxers use.

MediaInfo

Information about a piece of media as a whole.

SeekedTo

SeekedTo is the result of a seek.

Track

A Track is an independently coded media bitstream. A media format may contain multiple tracks in one container. Each of those tracks are represented by one Track.

TrackFlags

Flags indicating certain attributes about a track.

VendorDataAttachment

Application or vendor-specific proprietary binary data attachment.

Enums

Attachment

An attachment is additional data that is carried along with the container format.

SeekMode

SeekMode selects the precision of a seek.

SeekTo

SeekTo specifies a position to seek to.

TrackType

The track type.

Constants

FORMAT_ID_NULL

Null container format

Traits

FormatReader

A FormatReader is a media container demuxer. It provides methods to read a media container and iterate over the codec bitstream packets of all encapsulated tracks. Additionally, it provides methods to access any metadata, chapters, or attachments.