symphonia_core::formats

Module probe

The probe module provides methods and traits to support auto-detection of media formats from arbitrary media streams.

Structs

Hint

A Hint provides additional information and context when probing a media source stream.

Probe

Probe scans a MediaSourceStream for metadata and container formats, and provides an iterator-like interface to instantiate readers for the formats encountered.

ProbeDataMatchSpec

The probe match specification provides declarative information that is used by a Probe to detect the presence of a specific container or metadata format while scanning a MediaSourceStream.

ProbeFormatData

Container format-specific probe data.

ProbeMetadataData

Metadata format-specific probe data.

ProbeOptions

Options for controlling the behaviour of a Probe.

Enums

Anchors

Anchors are explicit fixed offsets from the end of a media stream that should be searched for format markers.

Score

The result of a scoring operation.

Traits

ProbeableFormat

To support probing, a FormatReader must implement the ProbeableFormat trait.

ProbeableMetadata

To support probing, a MetadataReader must implement the ProbeableMetadata trait.

Scoreable

The Scoreable trait defines the scoring functionality a reader must implement to support probing for a container or metadata format.

Type Aliases

FormatFactoryFn

FormatReader probe factory function. Creates a boxed FormatReader.

MetadataFactoryFn

MetadataReader probe factory function. Creates a boxed MetadataReader.