symphonia_core::formats::probe

Trait Scoreable

pub trait Scoreable {
    // Required method
    fn score(src: ScopedStream<&mut MediaSourceStream<'_>>) -> Result<Score>;
}

The Scoreable trait defines the scoring functionality a reader must implement to support probing for a container or metadata format.

Required Methods

fn score(src: ScopedStream<&mut MediaSourceStream<'_>>) -> Result<Score>

Using scoped access to a MediaSourceStream, calculate and return a value between 0 and 255 indicating the confidence of the reader in decoding or parsing the stream.

If the format is definitely not supported, then score should return Score::Unsupported since a score of 0 is still considered supported, even if unlikely.

If an error is returned, errors other than Error::IoError (excluding the unexpected EOF kind) are treated as if Score::Unsupported was returned. All other IO errors abort the probe operation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.