symphonia_core::formats

Struct Cue

pub struct Cue {
    pub index: u32,
    pub start_ts: u64,
    pub tags: Vec<Tag>,
    pub points: Vec<CuePoint>,
}

A Cue is a designated point of time within a media stream.

A Cue may be a mapping from either a source track, a chapter, cuesheet, or a timestamp depending on the source media. A Cueโs duration is the difference between the Cueโs timestamp and the next. Each Cue may contain an optional index of points relative to the Cue that never exceed the timestamp of the next Cue. A Cue may also have associated Tags.

Fields
index: u32

A unique index for the Cue.

start_ts: u64

The starting timestamp in number of frames from the start of the stream.

tags: Vec<Tag>

A list of Tags associated with the Cue.

points: Vec<CuePoint>

A list of CuePointss that are contained within this Cue. These points are children of the Cue since the Cue itself is an implicit CuePoint.