symphonia_core::formats

Struct Track

pub struct Track {
    pub id: u32,
    pub codec_params: Option<CodecParameters>,
    pub language: Option<String>,
    pub time_base: Option<TimeBase>,
    pub num_frames: Option<u64>,
    pub duration: Option<Duration>,
    pub start_ts: Timestamp,
    pub delay: Option<u32>,
    pub padding: Option<u32>,
    pub flags: TrackFlags,
}

A Track is an independently coded media bitstream. A media format may contain multiple tracks in one container. Each of those tracks are represented by one Track.

Fields
id: u32

A unique identifier for the track.

For most formats this is usually the zero-based index of the track, however, some more complex formats set this differently.

codec_params: Option<CodecParameters>

The codec parameters for the track.

If None, the format reader was unable to determine the codec parameters and the track will be unplayable.

language: Option<String>

The language of the track. May be unknown or not set.

time_base: Option<TimeBase>

The timebase of the track.

The timebase is the length of time in seconds of a single tick of a timestamp or duration. It can be used to convert any timestamp or duration related to the track into seconds.

num_frames: Option<u64>

The length of the track in number of audio, video, or subtitle frames.

This count excludes any delay or padding frames. In other words, it is the number of playable (non-discarded) frames.

Generally, when presenting a track’s duration, the duration field should be used instead. This is because a track’s timebase may not always be the reciprocal of the sample or frame rate, resulting in slight differences between the duration as stated in the container (in timebase units) vs. when calculated from the number of playable frames.

duration: Option<Duration>

The duration of the track in timebase units.

If a timebase is available, this field can be used to calculate the total duration of the track in seconds by using TimeBase::calc_time and passing the duration as the argument.

start_ts: Timestamp

The timestamp of the first frame.

delay: Option<u32>

The number of leading frames inserted by the encoder that should be skipped during playback.

padding: Option<u32>

The number of trailing frames inserted by the encoder for padding that should be skipped during playback.

flags: TrackFlags

Flags indicating track attributes.

Implementations

impl Track

pub fn new(id: u32) -> Self

Instantiate a new track with a given ID.

pub fn with_codec_params(&mut self, codec_params: CodecParameters) -> &mut Self

Provide the codec parameters.

Note: If the codec parameters contains a non-zero sample or frame rate that, a default timebase will be derived.

pub fn with_language(&mut self, language: &str) -> &mut Self

Provide the track language.

pub fn with_time_base(&mut self, time_base: TimeBase) -> &mut Self

Provide the TimeBase.

pub fn with_num_frames(&mut self, num_frames: u64) -> &mut Self

Provide the total number of frames.

pub fn with_duration(&mut self, duration: Duration) -> &mut Self

Provide the duration in timebase units.

pub fn with_start_ts(&mut self, start_ts: Timestamp) -> &mut Self

Provide the timestamp of the first frame.

pub fn with_delay(&mut self, delay: u32) -> &mut Self

Provide the number of delay frames.

pub fn with_padding(&mut self, padding: u32) -> &mut Self

Provide the number of padding frames.

pub fn with_flags(&mut self, flags: TrackFlags) -> &mut Self

Append provided track flags.

pub fn track_type(&self) -> Option<TrackType>

Get the track type.

Determining the track type requires knowing the codec parameters. If codec parameters is None, then this function will also return None.