symphonia_core::formats::util

Function trim_packet

pub fn trim_packet(packet: &mut Packet, delay: u32, num_frames: Option<u64>)

Given a Packet, the encoder delay in frames, and the number of non-delay or padding frames, adjust the packet’s timestamp and duration, and populate the trim information.