symphonia_core::io

Struct MediaSourceStream

pub struct MediaSourceStream { /* private fields */ }

MediaSourceStream is the main reader type for Symphonia.

By using type erasure and dynamic dispatch, MediaSourceStream wraps and hides the inner reader from the consumer, allowing any typical Reader to be used with Symphonia in a generic way, selectable at runtime.

MediaSourceStream is designed to provide speed and flexibility in a number of challenging I/O scenarios.

First, to minimize system call and dynamic dispatch overhead on the inner reader, and to amortize that overhead over many bytes, MediaSourceStream implements an exponentially growing read-ahead buffer. The read-ahead length starts at 1kB, and doubles in length as more sequential reads are performed until it reaches 32kB. Growing the read-ahead length over time reduces the excess data buffered on consecutive seek() calls.

Second, to better support non-seekable sources, MediaSourceStream implements a configurable length buffer cache. By default, the buffer caches allows backtracking by up-to the minimum of either buffer_len - 32kB or the total number of bytes read since instantiation or the last buffer cache invalidation. Note that regular a seek() will invalidate the buffer cache.

Implementations

impl MediaSourceStream

pub fn new( source: Box<dyn MediaSource>, options: MediaSourceStreamOptions, ) -> Self