symphonia_core::io

Struct MonitorStream

pub struct MonitorStream<B: ReadBytes, M: Monitor> { /* private fields */ }

A MonitorStream is a passive stream that observes all operations performed on the inner stream and forwards an immutable reference of the result to a Monitor.

Implementations

impl<B: ReadBytes, M: Monitor> MonitorStream<B, M>

pub fn new(inner: B, monitor: M) -> MonitorStream<B, M>

pub fn inner(&self) -> &B

pub fn inner_mut(&mut self) -> &mut B

pub fn into_inner(self) -> B

pub fn monitor(&self) -> &M

pub fn monitor_mut(&mut self) -> &mut M