symphonia_core::io

Struct ScopedStream

pub struct ScopedStream<B: ReadBytes> { /* private fields */ }

A ScopedStream restricts the number of bytes that may be read to an upper limit.

Implementations

impl<B: ReadBytes> ScopedStream<B>

pub fn new(inner: B, len: u64) -> Self

Instantiates a new ScopedStream with an upper limit on the number of bytes that can be read from the inner source.

pub fn inner(&self) -> &B

Returns an immutable reference to the inner stream.

pub fn inner_mut(&mut self) -> &mut B

Returns a mutable reference to the inner stream.

pub fn ignore(&mut self) -> Result<()>

Ignores the remainder of the ScopedStream.

pub fn into_inner(self) -> B

Convert the ScopedStream to the inner stream.