Trait symphonia_core::io::MediaSource

pub trait MediaSource: Read + Seek + Send + Sync {
    fn is_seekable(&self) -> bool;
    fn byte_len(&self) -> Option<u64>;
}

MediaSource is a composite trait of std::io::Read and std::io::Seek. A source must implement this trait to be used by MediaSourceStream.

Despite requiring the std::io::Seek trait, seeking is an optional capability that can be queried at runtime.

Required Methods

fn is_seekable(&self) -> bool

Returns if the source is seekable. This may be an expensive operation.

fn byte_len(&self) -> Option<u64>

Returns the length in bytes, if available. This may be an expensive operation.