symphonia_core::io
Module vlc

The vlc module provides support for decoding variable-length codes (VLC).

Structs

Codebook

Codebook is a variable-length code decoding table that may be used to efficiently read symbols from a source of bits.

CodebookBuilder

CodebookBuilder generates a Codebook using a provided codebook specification and description.

Entry8x8

Entry8x8 is a codebook entry for 8-bit values with codes up-to 8-bits.

Entry8x16

Entry8x16 is a codebook entry for 8-bit values with codes up-to 16-bits.

Entry8x32

Entry8x32 is a codebook entry for 8-bit values with codes up-to 32-bits.

Entry16x8

Entry16x8 is a codebook entry for 16-bit values with codes up-to 8-bits.

Entry16x16

Entry16x16 is a codebook entry for 16-bit values with codes up-to 16-bits.

Entry16x32

Entry16x32 is a codebook entry for 16-bit values with codes up-to 32-bits.

Entry32x8

Entry32x8 is a codebook entry for 32-bit values with codes up-to 8-bits.

Entry32x16

Entry32x16 is a codebook entry for 32-bit values with codes up-to 16-bits.

Entry32x32

Entry32x32 is a codebook entry for 32-bit values with codes up-to 32-bits.

Enums

BitOrder

BitOrder describes the relationship between the order of bits in the provided codewords and the order in which bits are read.

Traits

CodebookEntry

CodebookEntry provides the functions required for an entry in the Codebook.