symphonia_core

Module meta

The meta module defines basic metadata elements, and management structures.

Tags

Within the context of media, a tag is single piece of metadata about the media as a whole, or a track within the media. The storage of tags, their structure, and organization, varies based on the metadata/tagging format.

The Tag structure represents a single tag, and abstracts over the differences in tagging formats. Tag is a composition of a mandatory raw tag, and an optional standard tag.

Raw Tags

A RawTag stores a tag in a data format that matches as closely as possible to the format of the tag as it was written. The data format depends on the tagging format, and the writer of the tag.

A RawTag consists of a mandatory key-value pair. For most tagging formats, this is sufficient to faithfully represent the original tag, however, for some more structured tagging formats, a set of additional key-value pairs (RawTagSubField) may be populated.

The meaning of the tag can be derived from its key, however, the key may be named differently based on the underlying tagging format and the writer of the tag.

Raw tags can be ignored by most tag consumers. Instead, standard tags should be preferred.

Standard Tags

A StandardTag is a parsed representation of a tag. Unlike a raw tag, a standard tag has a well-defined data type and meaning.

A metadata reader will assign a StandardTag to a Tag if it is able to identify the meaning of the RawTag, and parse its value. If the RawTag maps to multiple StandardTags, then the Tag (along with the RawTag) will be duplicated for each StandardTag with each instance being assigned one StandardTag.

An end-user should prefer consuming standard tags over raw tags.

Storage Efficiency

In many cases, the value of a RawTag will be the same as the StandardTag. Since a value may be large, duplicating it could be wasteful. For this reason, string and binary data values are stored using an Arc.

Modules

well_known

IDs for well-known metadata formats.

Structs

Chapter

A chapter is a labelled section of a piece of media with a defined start time.

ChapterGroup

A group of chapters and/or other chapter groups.

ColorPaletteInfo

A description of the color palette for indexed color mode.

Metadata

A reference to the metadata inside of a MetadataLog.

MetadataBuffer

The decoded contents of read metadata.

MetadataBuilder

A builder for MetadataRevision.

MetadataContainer

A container of metadata tags and pictures.

MetadataId

A MetadataId is a unique identifier used to identify a specific metadata format.

MetadataInfo

Basic information about a metadata format.

MetadataLog

A queue for time-ordered MetadataRevisions.

MetadataOptions

A common set of options that all metadata readers use.

MetadataRevision

A metadata revision is a container for a single discrete version of media metadata.

PerTrackMetadata

Container for metadata associated with a specific track. A MetadataContainer wrapper associating it with a track ID.

PerTrackMetadataBuilder

A builder for PerTrackMetadata.

RawTag

A raw tag represents a tag in a data format that matches, as closely as possible, to the data format that the tag was written in.

RawTagSubField

A key-value pair of supplementary data that can be attached to a raw tag.

Size

A 2-dimensional (width and height) size type.

Tag

A tag encapsulates a single piece of metadata.

Visual

A Visual is any 2 dimensional graphic.

Enums

ChapterGroupItem

A chapter group item is either a chapter or chapter group.

ColorMode

Indicates how colors are represented in the image.

ColorModel

A color model describes how a color is represented.

ContentAdvisory

A content advisory.

MetadataSideData

Enumeration of types of metadata side data.

RawValue

The value of a RawTag.

StandardTag

A standard tag is an enumeration of well-defined and well-known tags with parsed values.

StandardVisualKey

StandardVisualKey is an enumeration providing standardized keys for common visual dispositions. A demuxer may assign a StandardVisualKey to a Visual if the disposition of the attached visual is known and can be mapped to a standard key.

Constants

METADATA_ID_NULL

Null metadata format

Traits

MetadataReader

A MetadataReader reads and decodes metadata and produces a revision of that decoded metadata.