symphonia_core::packet

Struct PacketBuilder

pub struct PacketBuilder<T, P, D, B> { /* private fields */ }

A builder for creating owning or non-owning packets.

See Packet for a detailed description of all packet fields.

The track ID, PTS, duration, and data fields are mandatory and must be provided before a packet can be built.

Implementations

impl PacketBuilder<NoTrackId, NoPts, NoDur, NoBuf>

pub fn new() -> Self

Create the packet builder.

impl PacketBuilder<HasTrackId, HasPts, HasDur, HasBuf>

pub fn build(self) -> Packet

Build the packet.

impl<'a> PacketBuilder<HasTrackId, HasPts, HasDur, HasBufRef<'a>>

pub fn build_packet_ref(self) -> PacketRef<'a>

Build a non-owning packet.

impl<T, B> PacketBuilder<T, HasPts, NoDur, B>

pub fn trimmed_dur( self, block_dur: Duration, end_pts: Option<Timestamp>, ) -> PacketBuilder<T, HasPts, HasDur, B>

Provide the packet’s duration and calculate the trim fields.

Given the provided duration including delay and padding frames, the stream end timestamp (optional), and the previously provided PTS, this function calculates the packet’s duration, trim start, and trim end.

This helper assumes that all frames with a PTS < 0 are to be trimmed from the start. Frames whose PTS exceeds the end timestamp of the stream are trimmed from the end. The trim end will only be calculated if the stream’s end timestamp is provided.

impl<T, P, B> PacketBuilder<T, P, NoDur, B>

pub fn dur(self, dur: Duration) -> PacketBuilder<T, P, HasDur, B>

Provide the packet’s duration including delay and padding frames.

impl<T, P, D, B> PacketBuilder<T, P, D, B>

pub fn track_id(self, track_id: u32) -> PacketBuilder<HasTrackId, P, D, B>

Provide the track ID.

pub fn pts(self, pts: Timestamp) -> PacketBuilder<T, HasPts, D, B>

Provide the presentation timestamp (PTS).

pub fn data(self, buf: impl Into<Box<[u8]>>) -> PacketBuilder<T, P, D, HasBuf>

Provide the packet’s data buffer.

When holding an owned data buffer, an owning Packet is built.

pub fn data_by_ref<'a>( self, buf: &'a [u8], ) -> PacketBuilder<T, P, D, HasBufRef<'a>>

Provide the packet’s data buffer as a non-owning reference.

When holding a non-owning data buffer reference, a non-owning PacketRef is built.

pub fn dts(self, dts: Timestamp) -> Self

Provide the decode timestamp (DTS).

pub fn trim_start(self, trim_start: Duration) -> Self

Provide the trim start duration.

pub fn trim_end(self, trim_end: Duration) -> Self

Provide the trim end duration.