Trait symphonia_core::probe::QueryDescriptor

pub trait QueryDescriptor {
    // Required methods
    fn query() -> &'static [Descriptor];
    fn score(context: &[u8]) -> u8;
}

The QueryDescriptor trait indicates that the implementer may be registered and capable of probing.

Required Methods

fn query() -> &'static [Descriptor]

Returns a list of descriptors.

fn score(context: &[u8]) -> u8

Using the provided context buffer, score calculate and returns a value between 0 and 255 indicating the confidence of the reader in decoding or parsing the source stream.

Object Safety

This trait is not object safe.