symphonia_core

Module sample

The sample module defines the core audio sample trait and any non-primitive sample data types.

Structs

i24

A signed 24-bit integer sample with an internal signed 32-bit integer representation.

u24

An unsigned 24-bit integer sample with an internal unsigned 32-bit integer representation.

Enums

SampleFormat

SampleFormat describes the data encoding for an audio sample.

Traits

Sample

Sample provides a common interface for manipulating sample’s regardless of the underlying data type. Additionally, Sample provides information regarding the format of underlying data types representing the sample when in memory, but also when exported.