symphonia_core::sample

Struct u24

pub struct u24(pub u32);

An unsigned 24-bit integer sample with an internal unsigned 32-bit integer representation.

There are no guarantees the sample is within the valid range 24-bit range. Use the Sample::clamped function to clamp the sample to the valid range.

Tuple Fields
0: u32
Implementations

impl u24

pub const MAX: u24

The largest value that can be represented by this integer type.

pub const MIN: u24

The smallest value that can be represented by this integer type.

pub fn into_u32(self) -> u32

๐Deprecated: Superseded by inner.

Get the underlying u32 backing this u24.

pub fn inner(self) -> u32

Get the underlying u32 backing this u24.

pub fn to_ne_bytes(self) -> [u8; 3]

Return the memory representation of this u24 as a byte array in native byte order.