Module symphonia_core::units

The units module provides definitions for common units.

Structs

Time

Time represents a duration of time in seconds, or the number of seconds since an arbitrary epoch. Time is stored as an integer number of seconds plus any remaining fraction of a second as a floating point value.

TimeBase

A TimeBase is the conversion factor between time, expressed in seconds, and a TimeStamp or Duration.

Type Aliases

Duration

A Duration indicates a positive span of time.

TimeStamp

A TimeStamp represents an instantenous instant in time since the start of a stream. One TimeStamp “tick” is equivalent to the stream’s TimeBase in seconds.