symphonia_core::audio

Struct AudioBuffer

pub struct AudioBuffer<S: Sample> { /* private fields */ }

AudioBuffer is a container for multi-channel planar audio sample data. An AudioBuffer is characterized by the duration (capacity), and audio specification (channels and sample rate). The capacity of an AudioBuffer is the maximum number of samples the buffer may store per channel. Manipulation of samples is accomplished through the Signal trait or direct buffer manipulation.

Implementations

impl<S: Sample> AudioBuffer<S>

pub fn new(duration: Duration, spec: SignalSpec) -> Self

Instantiate a new AudioBuffer using the specified signal specification and of the given duration.

pub fn unused() -> Self

Instantiates an unused AudioBuffer. An unused AudioBuffer will not allocate any memory, has a sample rate of 0, and no audio channels.

pub fn is_unused(&self) -> bool

Returns true if the AudioBuffer is unused.

pub fn spec(&self) -> &SignalSpec

Gets the signal specification for the buffer.

pub fn capacity(&self) -> usize

Gets the total capacity of the buffer. The capacity is the maximum number of audio frames a buffer can store.

pub fn planes(&self) -> AudioPlanes<'_, S>

Gets immutable references to all audio planes (channels) within the audio buffer.

Note: This is not a cheap operation for audio buffers with > 8 channels. It is advisable that this call is only used when operating on large batches of frames. Generally speaking, it is almost always better to use chan() to selectively choose the plane to read instead.

pub fn planes_mut(&mut self) -> AudioPlanesMut<'_, S>

Gets mutable references to all audio planes (channels) within the buffer.

Note: This is not a cheap operation for audio buffers with > 8 channels. It is advisable that this call is only used when modifying large batches of frames. Generally speaking, it is almost always better to use render(), fill(), chan_mut(), and chan_pair_mut() to modify the buffer instead.

pub fn convert<T: Sample>(&self, dest: &mut AudioBuffer<T>)
where S: IntoSample<T>,

Converts the contents of an AudioBuffer into an equivalent destination AudioBuffer of a different type. If the types are the same then this is a copy operation.

pub fn make_equivalent<E: Sample>(&self) -> AudioBuffer<E>

Makes an equivalent AudioBuffer of a different type.