Trait symphonia_core::audio::RawSample

pub trait RawSample: Sample {
    type RawType: Copy + Default + Sealed;

    fn into_raw_sample(self) -> Self::RawType;
}

RawSample provides a typed interface for converting a Sample from it’s in-memory data type to actual binary type.

Required Associated Types

type RawType: Copy + Default + Sealed

The RawType is a primitive data type, or fixed-size byte array, that is the final binary representation of the sample when written out to a byte-buffer.

Required Methods

fn into_raw_sample(self) -> Self::RawType