Module symphonia_core::errors

The errors module defines the common error type.

Enums

Error

Error provides an enumeration of all possible errors reported by Symphonia.

SeekErrorKind

SeekErrorKind is a list of generic reasons why a seek may fail.

Functions

decode_error

Convenience function to create a decode error.

end_of_stream_error

Convenience function to create an end-of-stream error.

limit_error

Convenience function to create a limit error.

reset_error

Convenience function to create a reset required error.

seek_error

Convenience function to create a seek error.

unsupported_error

Convenience function to create an unsupport feature error.

Type Definitions

Result