Module symphonia_core::formats

The format module provides the traits and support structures necessary to implement media demuxers.

Modules

prelude

The formats module prelude.

util

Helper utilities for implementing FormatReaders.

Structs

Cue

A Cue is a designated point of time within a media stream.

CuePoint

A CuePoint is a point, represented as a frame offset, within a Cue.

FormatOptions

FormatOptions is a common set of options that all demuxers use.

Packet

A Packet contains a discrete amount of encoded data for a single codec bitstream. The exact amount of data is bounded, but not defined, and is dependant on the container and/or the encapsulated codec.

SeekedTo

SeekedTo is the result of a seek.

Track

A Track is an independently coded media bitstream. A media format may contain multiple tracks in one container. Each of those tracks are represented by one Track.

Enums

SeekMode

SeekMode selects the precision of a seek.

SeekTo

SeekTo specifies a position to seek to.

Traits

FormatReader

A FormatReader is a container demuxer. It provides methods to probe a media container for information and access the tracks encapsulated in the container.