symphonia_core::formats

Module util

Helper utilities for implementing FormatReaders.

Structs

SeekIndex

A SeekIndex stores SeekPoints (generally a sample or frame number to byte offset) within a media stream and provides methods to efficiently search for the nearest SeekPoint(s) given a timestamp.

SeekPoint

A SeekPoint is a mapping between a sample or frame number to byte offset within a media stream.

Enums

SeekSearchResult

SeekSearchResult is the return value for a search on a SeekIndex. It returns a range of SeekPoints a FormatReader should search to find the desired timestamp. Ranges are lower-bound inclusive, and upper-bound exclusive.

Functions

trim_packet

Given a Packet, the encoder delay in frames, and the number of non-delay or padding frames, adjust the packet’s timestamp and duration, and populate the trim information.