symphonia_core::io

Trait FiniteStream

pub trait FiniteStream {
    // Required methods
    fn byte_len(&self) -> u64;
    fn bytes_read(&self) -> u64;
    fn bytes_available(&self) -> u64;
}

A FiniteStream is a stream that has a known length in bytes.

Required Methods

fn byte_len(&self) -> u64

Returns the length of the the stream in bytes.

fn bytes_read(&self) -> u64

Returns the number of bytes that have been read.

fn bytes_available(&self) -> u64

Returns the number of bytes available for reading.