Trait symphonia_core::io::ReadBitsRtl

pub trait ReadBitsRtl: FetchBitsRtl {

    fn realign(&mut self) { ... }
    fn ignore_bits(&mut self, num_bits: u32) -> Result<()> { ... }
    fn ignore_bit(&mut self) -> Result<()> { ... }
    fn read_bool(&mut self) -> Result<bool> { ... }
    fn read_bit(&mut self) -> Result<u32> { ... }
    fn read_bits_leq32(&mut self, bit_width: u32) -> Result<u32> { ... }
    fn read_bits_leq32_signed(&mut self, bit_width: u32) -> Result<i32> { ... }
    fn read_bits_leq64(&mut self, bit_width: u32) -> Result<u64> { ... }
    fn read_bits_leq64_signed(&mut self, bit_width: u32) -> Result<i64> { ... }
    fn read_unary_zeros(&mut self) -> Result<u32> { ... }
    fn read_unary_zeros_capped(&mut self, limit: u32) -> Result<u32> { ... }
    fn read_unary_ones(&mut self) -> Result<u32> { ... }
    fn read_unary_ones_capped(&mut self, limit: u32) -> Result<u32> { ... }
    fn read_codebook<E: CodebookEntry>(
        &mut self,
        codebook: &Codebook<E>
    ) -> Result<(E::ValueType, u32)> { ... }

}

ReadBitsRtl reads bits from least-significant to most-significant.

Provided Methods

fn realign(&mut self)

Discards any saved bits and resets the BitStream to prepare it for a byte-aligned read.

fn ignore_bits(&mut self, num_bits: u32) -> Result<()>

Ignores the specified number of bits from the stream or returns an error.

fn ignore_bit(&mut self) -> Result<()>

Ignores one bit from the stream or returns an error.

fn read_bool(&mut self) -> Result<bool>

Read a single bit as a boolean value or returns an error.

fn read_bit(&mut self) -> Result<u32>

Reads and returns a single bit or returns an error.

fn read_bits_leq32(&mut self, bit_width: u32) -> Result<u32>

Reads and returns up to 32-bits or returns an error.

fn read_bits_leq32_signed(&mut self, bit_width: u32) -> Result<i32>

Reads up to 32-bits and interprets them as a signed two’s complement integer or returns an error.

fn read_bits_leq64(&mut self, bit_width: u32) -> Result<u64>

Reads and returns up to 64-bits or returns an error.

fn read_bits_leq64_signed(&mut self, bit_width: u32) -> Result<i64>

Reads up to 64-bits and interprets them as a signed two’s complement integer or returns an error.

fn read_unary_zeros(&mut self) -> Result<u32>

Reads and returns a unary zeros encoded integer or an error.

fn read_unary_zeros_capped(&mut self, limit: u32) -> Result<u32>

Reads and returns a unary zeros encoded integer that is capped to a maximum value.

fn read_unary_ones(&mut self) -> Result<u32>

Reads and returns a unary ones encoded integer or an error.

fn read_unary_ones_capped(&mut self, limit: u32) -> Result<u32>

Reads and returns a unary ones encoded integer or an error.

fn read_codebook<E: CodebookEntry>(
&mut self,
codebook: &Codebook<E>
) -> Result<(E::ValueType, u32)>

Reads a codebook value from the BitStream using the provided Codebook and returns the decoded value or an error.