symphonia_core::meta

Enum StandardTagKey

pub enum StandardTagKey {

    AcoustidFingerprint,
    AcoustidId,
    Album,
    AlbumArtist,
    Arranger,
    Artist,
    Bpm,
    Comment,
    Compilation,
    Composer,
    Conductor,
    ContentGroup,
    Copyright,
    Date,
    Description,
    DiscNumber,
    DiscSubtitle,
    DiscTotal,
    EncodedBy,
    Encoder,
    EncoderSettings,
    EncodingDate,
    Engineer,
    Ensemble,
    Genre,
    IdentAsin,
    IdentBarcode,
    IdentCatalogNumber,
    IdentEanUpn,
    IdentIsrc,
    IdentPn,
    IdentPodcast,
    IdentUpc,
    Label,
    Language,
    License,
    Lyricist,
    Lyrics,
    MediaFormat,
    MixDj,
    MixEngineer,
    Mood,
    MovementName,
    MovementNumber,
    MusicBrainzAlbumArtistId,
    MusicBrainzAlbumId,
    MusicBrainzArtistId,
    MusicBrainzDiscId,
    MusicBrainzGenreId,
    MusicBrainzLabelId,
    MusicBrainzOriginalAlbumId,
    MusicBrainzOriginalArtistId,
    MusicBrainzRecordingId,
    MusicBrainzReleaseGroupId,
    MusicBrainzReleaseStatus,
    MusicBrainzReleaseTrackId,
    MusicBrainzReleaseType,
    MusicBrainzTrackId,
    MusicBrainzWorkId,
    Opus,
    OriginalAlbum,
    OriginalArtist,
    OriginalDate,
    OriginalFile,
    OriginalWriter,
    Owner,
    Part,
    PartTotal,
    Performer,
    Podcast,
    PodcastCategory,
    PodcastDescription,
    PodcastKeywords,
    Producer,
    PurchaseDate,
    Rating,
    ReleaseCountry,
    ReleaseDate,
    Remixer,
    ReplayGainAlbumGain,
    ReplayGainAlbumPeak,
    ReplayGainTrackGain,
    ReplayGainTrackPeak,
    Script,
    SortAlbum,
    SortAlbumArtist,
    SortArtist,
    SortComposer,
    SortTrackTitle,
    TaggingDate,
    TrackNumber,
    TrackSubtitle,
    TrackTitle,
    TrackTotal,
    TvEpisode,
    TvEpisodeTitle,
    TvNetwork,
    TvSeason,
    TvShowTitle,
    Url,
    UrlArtist,
    UrlCopyright,
    UrlInternetRadio,
    UrlLabel,
    UrlOfficial,
    UrlPayment,
    UrlPodcast,
    UrlPurchase,
    UrlSource,
    Version,
    Writer,

}

StandardTagKey is an enumeration providing standardized keys for common tag types. A tag reader may assign a StandardTagKey to a Tag if the tag’s key is generally accepted to map to a specific usage.

Variants

AcoustidFingerprint

AcoustidId

Album

AlbumArtist

Arranger

Artist

Bpm

Comment

Compilation

Composer

Conductor

ContentGroup

Copyright

Date

Description

DiscNumber

DiscSubtitle

DiscTotal

EncodedBy

Encoder

EncoderSettings

EncodingDate

Engineer

Ensemble

Genre

IdentAsin

IdentBarcode

IdentCatalogNumber

IdentEanUpn

IdentIsrc

IdentPn

IdentPodcast

IdentUpc

Label

Language

License

Lyricist

Lyrics

MediaFormat

MixDj

MixEngineer

Mood

MovementName

MovementNumber

MusicBrainzAlbumArtistId

MusicBrainzAlbumId

MusicBrainzArtistId

MusicBrainzDiscId

MusicBrainzGenreId

MusicBrainzLabelId

MusicBrainzOriginalAlbumId

MusicBrainzOriginalArtistId

MusicBrainzRecordingId

MusicBrainzReleaseGroupId

MusicBrainzReleaseStatus

MusicBrainzReleaseTrackId

MusicBrainzReleaseType

MusicBrainzTrackId

MusicBrainzWorkId

Opus

OriginalAlbum

OriginalArtist

OriginalDate

OriginalFile

OriginalWriter

Owner

Part

PartTotal

Performer

Podcast

PodcastCategory

PodcastDescription

PodcastKeywords

Producer

PurchaseDate

Rating

ReleaseCountry

ReleaseDate

Remixer

ReplayGainAlbumGain

ReplayGainAlbumPeak

ReplayGainTrackGain

ReplayGainTrackPeak

Script

SortAlbum

SortAlbumArtist

SortArtist

SortComposer

SortTrackTitle

TaggingDate

TrackNumber

TrackSubtitle

TrackTitle

TrackTotal

TvEpisode

TvEpisodeTitle

TvNetwork

TvSeason

TvShowTitle

Url

UrlArtist

UrlCopyright

UrlInternetRadio

UrlLabel

UrlOfficial

UrlPayment

UrlPodcast

UrlPurchase

UrlSource

Version

Writer