symphonia_core

Module probe

The probe module provides methods and traits to support auto-detection of media formats from arbitrary media streams.

Structs

Descriptor

Descriptor provides declarative information about container and metadata formats. Descriptors are used by Probe and related machinery to scan a MediaSourceStream for media.

Hint

A Hint provides additional information and context when probing a media source stream.

Probe

Probe scans a MediaSourceStream for metadata and container formats, and provides an iterator-like interface to instantiate readers for the formats encountered.

ProbeResult

ProbeResult contains the result of a format probe operation.

ProbedMetadata

Metadata that came from the metadata field of ProbeResult.

Enums

Instantiate

Instantiate is an enumeration of instantiation functions used by Descriptor and Probe to instantiate metadata and container format readers.

Traits

QueryDescriptor

The QueryDescriptor trait indicates that the implementer may be registered and capable of probing.